inspecao_fiscalizacaoapp/requirements.txt
python-decouple
oracledb
djangorestframework
markdown
django-filter
django
pillow
django-cors-headers
